+++ scenes/manager.gd
@@ -12,19 +12,26 @@ var layers: Dictionary = {}
 func _ready():
 	%Neurons.set_right_disconnects(true)
 
-func get_inputs() -> void:
+func _show_error() -> void:
+	%Popup.title = "Validation Error"
+	%Popup.dialog_text = "Not all input fields are valid"
+	%Popup.popup_centered()
+
+func get_network_inputs() -> void:
 	inputs_count = max(int(%InputsCount.value), 0)
-	learning_rate = min(float(%LearningRate.value), 1.0)
 	layers_count = int(%LayersCount.value)
-	epoch = max(int(%Epoch.value), 0)
 	activation_function = %ActivationFunction.value
 	%InputsCount.value = str(inputs_count)
-	%LearningRate.value = str(learning_rate)
 	%LayersCount.value = str(layers_count)
+
+func get_learning_inputs() -> void:
+	learning_rate = min(float(%LearningRate.value), 1.0)
+	epoch = max(int(%Epoch.value), 0)
+	%LearningRate.value = str(learning_rate)
 	%Epoch.value = str(epoch)
 
-func are_inputs_valid() -> bool:
-	if inputs_count > 0 && learning_rate > 0 && layers_count > 0 && epoch > 0 && activation_function != "None":
+func are_network_inputs_valid() -> bool:
+	if inputs_count > 0 && layers_count > 0 && activation_function != "":
 		var all_valid = true
 		for i in range(%LayersContainer.get_child_count()):
 			var node = %LayersContainer.get_child(i)
@@ -33,15 +40,37 @@ func are_inputs_valid() -> bool:
 				all_valid = false
 		if all_valid:
 			return true
-	%Popup.title = "Validation Error"
-	%Popup.dialog_text = "Not all input fields are valid"
-	%Popup.popup_centered()
+	_show_error()
+	return false
+
+func are_learning_inputs_valid() -> bool:
+	if learning_rate > 0 && learning_rate <= 1 && epoch >= 1:
+		return true
+	_show_error()
 	return false
 
 func _on_apply_changes_button_up():
-	get_inputs()
-	if not are_inputs_valid():
+	get_network_inputs()
+	if not are_network_inputs_valid():
 		return
+
+	var prev = ""
+	for key in layers:
+		if prev == "":
+			prev = key
+			continue
+		if key == "out":
+			for i in range(len(layers[prev].all_neurons)):
+				%Neurons.disconnect_node(prev, i, key, i)
+		elif prev == "in":
+			for i in range(len(layers[key].all_neurons)):
+				for j in range(len(layers[prev].inputs_outputs)):
+					%Neurons.disconnect_node(prev, j, key, i)
+		else:
+			for i in range(len(layers[prev].all_neurons)):
+				for j in range(len(layers[key].all_neurons)):
+					%Neurons.disconnect_node(prev, i, key, j)
+		prev = key
 
 	# Clean All
 	for i in range(%Neurons.get_child_count()):
@@ -74,8 +103,8 @@ func _on_apply_changes_button_up():
 	outputs_layer.inputs_outputs_count = nc_of_each_layer[len(nc_of_each_layer) - 1]
 	outputs_layer.position_offset.x = (layers_count + 1) * 600
 	layers["out"] = outputs_layer
-	outputs_layer.ready.connect(_connect_all_lines)
 	%Neurons.add_child(outputs_layer)
+	%ConnectionsButton.disabled = false
 
 func _on_layers_count_text_changed(inputControl: InputControl):
 	var count = int(%LayersCount.value)
@@ -129,9 +158,41 @@ func _on_neurons_disconnection_request(from_node, from_port, to_node, to_port):
 	if disconnect:
 		%Neurons.disconnect_node(from_node, from_port, to_node, to_port)
 
-func _connect_all_lines() -> void:
-	var prev = null
-	for layer in layers:
-		if prev == null:
-			prev = layer
+func _on_train_button_button_up():
+	get_learning_inputs()
+	if not are_learning_inputs_valid():
+		return
+
+func _on_test_button_button_up():
+	get_learning_inputs()
+	if not are_learning_inputs_valid():
+		return
+
+func _on_connections_button_button_up():
+	var prev = ""
+	for key in layers:
+		if prev == "":
+			prev = key
 			continue
+		if key == "out":
+			for i in range(len(layers[prev].all_neurons)):
+				layers[prev].all_neurons[i].connect_output(layers[key].inputs_outputs[i])
+				%Neurons.connect_node(prev, i, key, i)
+		elif prev == "in":
+			for i in range(len(layers[key].all_neurons)):
+				for j in range(len(layers[prev].inputs_outputs)):
+					layers[key].all_neurons[i].connect_input("W in-" + str(j + 1), layers[prev].inputs_outputs[j])
+					%Neurons.connect_node(prev, j, key, i)
+		else:
+			for i in range(len(layers[prev].all_neurons)):
+				for j in range(len(layers[key].all_neurons)):
+					layers[key].all_neurons[j].connect_neuron("W " + str(layers[prev].layer_id) + "-" + str(i + 1), layers[prev].all_neurons[i])
+					%Neurons.connect_node(prev, i, key, j)
+		prev = key
+	%ConnectionsButton.disabled = true
+
+func _on_feed_forward_backward_button_button_up():
+	for key in layers:
+		if key == "in" or key == "out":
+			continue
+		layers[key].feed_forward_all_nodes()